+++ src/common/client.ts
@@ -525,7 +525,7 @@ export class Client extends EventEmitter {
         this._input = new Input(this);
         this._input.on('scroll-lock', (lock) => {
             this.display.scrollLock = lock;
-            this.display.scrollDisplay();
+            this.display.scrollDisplay(true);
             this.emit('scroll-lock', lock);
         });